refactor: account daily 관련 타입 declare 컨버팅
