+++ src/components/account/AccountBookDailyCard.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { AccountBookDailySum, AccountBookDailyItem } from '@components/account';
-import type { SingleAccount } from '@hooks/account/useDailyAccount';
 
 interface SingleAccountProp {
   items: {
